+++ lib/boruta/oauth/requests/hybrid_request.ex
@@ -10,26 +10,28 @@ defmodule Boruta.Oauth.HybridRequest do
   """
   @type t :: %__MODULE__{
           client_id: String.t(),
-          redirect_uri: String.t(),
-          state: String.t(),
-          nonce: String.t(),
-          scope: String.t(),
-          resource_owner: struct(),
-          grant_type: String.t(),
           code_challenge: String.t(),
           code_challenge_method: String.t(),
-          response_types: list(String.t())
+          grant_type: String.t(),
+          nonce: String.t(),
+          redirect_uri: String.t(),
+          resource_owner: struct(),
+          response_mode: String.t(),
+          response_types: list(String.t()),
+          scope: String.t(),
+          state: String.t()
         }
   @enforce_keys [:client_id, :redirect_uri, :resource_owner]
   defstruct client_id: nil,
-            redirect_uri: nil,
-            state: "",
-            nonce: "",
-            scope: "",
-            resource_owner: nil,
-            grant_type: "authorization_code",
             code_challenge: "",
             code_challenge_method: "plain",
+            grant_type: "authorization_code",
+            nonce: "",
+            prompt: "",
+            redirect_uri: nil,
+            resource_owner: nil,
+            response_mode: "fragment",
             response_types: [],
-            prompt: ""
+            scope: "",
+            state: ""
 end